openfaas/generate_password/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.3
pillow==9.0.0
cryptography==36.0.1
